Deletion logic added
